develop update 文件上传测试成功，需要再次完善

+++ autonet-api/src/main/java/cn/xiaoxige/autonet_api/presenter/AutoNetPresenter.java
@@ -366,9 +366,9 @@ public class AutoNetPresenter {
     public void doPushPost() {
         DoPushStreamPostUsecase usecase = new DoPushStreamPostUsecase(mRepo, mRequestEntity, mMediaType, mFileKey, mFile);
         final AAutoNetStreamCallback callback = (AAutoNetStreamCallback) mCallback;
-        usecase.execute(new DefaultSubscriber<Integer>() {
+        usecase.execute(new DefaultSubscriber<Float>() {
             @Override
-            public void DefaultOnNext(Integer data) {
+            public void DefaultOnNext(Float data) {
                 super.DefaultOnNext(data);
                 if (callback != null) {
                     callback.onPregress(data);
